feat(basket): added add and remove basket

+++ src/app/shared/services/basket.service.ts
@@ -1,11 +1,9 @@
 import {Injectable} from '@angular/core';
-import {BehaviorSubject} from "rxjs";
 
 @Injectable({
   providedIn: 'root'
 })
 export class BasketService {
-  public basket$?: BehaviorSubject<[]>;
 
   constructor() {
 
@@ -18,16 +16,31 @@ export class BasketService {
     return [];
   }
 
-  public setBasket(id: string) {
-    let newBasketItem = this.getBasket;
-    newBasketItem.push(id);
-    localStorage.setItem('basket', JSON.stringify(newBasketItem));
+  public setBasket(id: string): void {
+    if (!this.isAlreadyBasketItem(id)) {
+      let newBasketItem = this.getBasket;
+      newBasketItem.push(id);
+      localStorage.setItem('basket', JSON.stringify(newBasketItem));
+    }
+  }
+
+  public removeBasketItem(id: string): void {
+    let newBasketItem = this.getBasket,
+      alreadyBasketItemIndex = newBasketItem.findIndex((item: string) => item === id);
+    if (alreadyBasketItemIndex !== null) {
+      newBasketItem.splice(alreadyBasketItemIndex, 1);
+      localStorage.setItem('basket', JSON.stringify(newBasketItem));
+    }
+  }
+
+  public removeAllBasket(): void {
+    localStorage.removeItem('basket');
   }
 
   public isAlreadyBasketItem(id: string): boolean {
     if (!this.getBasket) {
       return false;
     }
-    return Boolean(this.getBasket.find((item: { id: string; }) => item.id === id));
+    return Boolean(this.getBasket.find((item: string) => item === id));
   }
 }
